Move element/jax.js to element/mml/jax.js (missed it on first round of renaming).
